fix: mobile styling of community & speaking section
